+++ oxen-rust/src/lib/src/api/client.rs
@@ -20,7 +20,9 @@ pub mod data_frames;
 pub mod diff;
 pub mod dir;
 pub mod entries;
+pub mod export;
 pub mod file;
+pub mod import;
 pub mod merger;
 pub mod metadata;
 pub mod oxen_version;
